LPS-199770 Move utils to utils folder and organize functions in different files
