+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/objectFolderContext.tsx
@@ -13,7 +13,7 @@ import {
 	TAction,
 	TState,
 } from '../types';
-import {getObjectFolderName} from '../utils';
+import {getObjectFolderName} from '../utils/diagram';
 import {ObjectFolderReducer} from './objectFolderReducer';
 
 interface ObjectFolderContextProps extends Array<TState | Function> {
