+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/objectFolderReducer.ts
@@ -16,7 +16,7 @@ import {
 	TAction,
 	TState,
 } from '../types';
-import {updateURLParam} from '../utils';
+import {updateURLParam} from '../utils/url';
 import {
 	convertAllObjectFieldsToUnselected,
 	getNonOverlappingEdges,
