+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/utils/diagram.d.ts
@@ -43,7 +43,3 @@ export declare function getEdgeParams(
 	targetY: number;
 };
 export declare function getObjectFolderName(): string;
-export declare function updateURLParam(
-	paramType: string,
-	paramValue: string
-): void;
